tools/requirements.txt
ruamel.yaml
pyserial
elf-size-analyze
